(chore) Readme has link to Dachshund repo
This repo is meant to have generic deployment instructions,
so that anyone can reuse it and deploy their own images.

However, new joiners such as myself were confused when
looking for instructions on how to deploy the specific dxw
dashboard.

This commit adds a link to the private repo where there are
dxw specific instructions.