(feature) Ensure page gets zoomed-to-fit on load
Ported over from Leeky’s old merge request[1]:

In certain situations FitJS wouldn't be triggered properly when the site
first loads, which caused the dashboard to fail to fit on the screen
properly.

Triggering FitJS manually, once after page load, and another one second
later for good measure, should ensure this always happens.

Additionally, I've removed the event listener for "DOMContentLoaded" as
soon as it is triggered, so there's one less thing for the Pi's browser
to have to do.

[1] https://git.govpress.com/dxw/scheduling-dashboard/merge_requests/24/diffs