(fix) Reconfigure logging for Dalmatian
We have recently deployed this app to Dalmatian for the first time.
Chris noticed that there are no request logs for the Dalmatian app in
Papertrail.

In production, we want the app to log to the container's output so that
the log output can be collected and uploaded to Papertrail.

Currently, the app only logs to STDOUT if RAILS_LOG_TO_STDOUT is set in
the environment, which it is not. Since we _always_ want to log to
STDOUT in production, just remove this check.